requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
black==20.8b1
click==7.1.2
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0

pluggy==0.13.1
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
zipp==3.1.0
